Move K8sClientConfig from argo/common to common
