+++ provider-service/base/go.mod
@@ -13,6 +13,7 @@ require (
 	github.com/onsi/gomega v1.27.7
 	github.com/sky-uk/kfp-operator v0.6.0
 	github.com/sky-uk/kfp-operator/argo/common v0.0.0-20250127163801-188c8090ec65
+	github.com/sky-uk/kfp-operator/common v0.0.0-00010101000000-000000000000
 	github.com/spf13/viper v1.19.0
 	github.com/stretchr/testify v1.10.0
 	go.uber.org/zap v1.24.0
@@ -21,7 +22,10 @@ require (
 	sigs.k8s.io/controller-runtime v0.15.0
 )
 
-replace github.com/sky-uk/kfp-operator => ../..
+replace (
+	github.com/sky-uk/kfp-operator => ../..
+	github.com/sky-uk/kfp-operator/common => ../../common
+)
 
 require (
 	cloud.google.com/go v0.112.1 // indirect
@@ -101,7 +105,7 @@ require (
 	golang.org/x/term v0.22.0 // indirect
 	golang.org/x/text v0.16.0 // indirect
 	golang.org/x/time v0.6.0 // indirect
-	golang.org/x/tools v0.21.1-0.20240508182429-e35e4ccd0d2d // indirect
+	golang.org/x/tools v0.22.0 // indirect
 	gomodules.xyz/jsonpatch/v2 v2.3.0 // indirect
 	google.golang.org/api v0.171.0 // indirect
 	google.golang.org/genproto v0.0.0-20240213162025-012b6fc9bca9 // indirect
